requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
openai>=1.50.0
python-dotenv>=1.0.0
pydantic>=2.9.0
jinja2>=3.1.4
python-multipart>=0.0.12
